✨ Add MyScoreBox component
