+++ src/Page/QuestionDetailPage/Short/style.css.ts
@@ -108,24 +108,3 @@ export const resultAnswerStyle = styleVariants({
     },
   ],
 });
-
-export const scoreStyle = style({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'end',
-  gap: '.5rem',
-});
-
-export const scoreLabelStyle = style({
-  fontWeight: '400',
-  fontSize: '1.125rem',
-  lineHeight: '1.625rem',
-  color: themeColors.text[9],
-});
-
-export const scoreValueStyle = style({
-  fontWeight: '700',
-  fontSize: '1.5rem',
-  lineHeight: '2.1875rem',
-  color: COLOR.RED,
-});
